# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Customer service chat system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/{id}:
    get:
      operationId: getPublicAgent
      tags: [agent]
      summary: Get reception agent by ID for dedicated chat link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent public profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentPublic"
        "404":
          description: Agent not found or not available
  /agent/login:
    post:
      operationId: agentLogin
      tags: [agent]
      summary: Agent login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentSession"
        "401":
          description: Invalid credentials
  /agent/me:
    get:
      operationId: getAgentMe
      tags: [agent]
      summary: Get current agent info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Agent info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentInfo"
        "401":
          description: Not authenticated
    patch:
      operationId: updateAgentMe
      tags: [agent]
      summary: Update current agent profile (display name and avatar)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentMeBody"
      responses:
        "200":
          description: Updated agent profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentInfo"
        "400":
          description: Validation error
        "401":
          description: Not authenticated
  /agent/me/password:
    patch:
      operationId: changeAgentPassword
      tags: [agent]
      summary: Change current agent password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                  minLength: 1
                newPassword:
                  type: string
                  minLength: 6
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        "400":
          description: Validation error
        "401":
          description: Not authenticated
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AgentCredentials:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AgentSession:
      type: object
      required: [token, agentId, userId, username, role]
      properties:
        token:
          type: string
        agentId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [agent, super_admin]
    AgentInfo:
      type: object
      required: [agentId, userId, username, role, displayName]
      properties:
        agentId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        role:
          type: string
          enum: [agent, super_admin]
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
    UpdateAgentMeBody:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
        avatarUrl:
          type: ["string", "null"]
    AgentPublic:
      type: object
      required: [id, displayName, isOnline]
      properties:
        id:
          type: integer
        displayName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        introduction:
          type: ["string", "null"]
        isOnline:
          type: boolean
...